app/requirements.txt
python-craigslist
python-dateutil
slackclient
sqlalchemy
